+++ conf/script/src/build_system/cmd/clean_build_dir/cli.py
@@ -6,7 +6,7 @@ import utils.cli.arg_parsing
 import utils.error.cls_def
 import utils.error.managed
 from build_system import cmd
-from build_system.cmd.hierarchy.find_build_dir.find_build_dir import BUILD_DIR_NAME
+from build_system.cmd.hierarchy.find_build_dir.impl import BUILD_DIR_NAME
 from utils.more_typing import AnyPath
 
 ROOT_DIR_ARG: Final[utils.cli.arg.CLIArg] = utils.cli.arg.CLIArg('rootdir')
