+++ conf/script/src/build_system/cmd/hierarchy/find_build_dir/__init__.py
@@ -1,1 +1,1 @@
-from build_system.cmd.hierarchy.find_build_dir.find_build_dir import find_build_dir_path, find_build_dir, BUILD_DIR_NAME
+from build_system.cmd.hierarchy.find_build_dir.impl import find_build_dir_path, find_build_dir, BUILD_DIR_NAME
